FINAL REFACTORING (DEBUGGING NEEDED)
