+++ src/features/projects/components/index.js
@@ -1,0 +1,4 @@
+export { default as ProjectCard } from './ProjectCard.desktop';
+export { default as ProjectCardMobile } from './ProjectCard.mobile';
+export { default as ProjectModal } from './ProjectModal.desktop';
+export { default as ProjectModalMobile } from './ProjectModal.mobile';
