+++ src/features/projects/index.js
@@ -1,4 +1,4 @@
-import { useIsMobile } from '@hooks/useIsMobile';
+import { useIsMobile } from '@/hooks/useMedia';
 import ProjectsDesktop from './Projects.desktop';
 import ProjectsMobile from './Projects.mobile';
 
@@ -6,3 +6,5 @@ export default function Projects() {
   const isMobile = useIsMobile();
   return isMobile ? <ProjectsMobile /> : <ProjectsDesktop />;
 }
+
+export * from './components';
