+++ src/features/skills/components/index.js
@@ -1,0 +1,4 @@
+export { default as SkillCard } from './SkillCard.desktop';
+export { default as SkillCardMobile } from './SkillCard.mobile';
+export { default as SkillCategory } from './SkillCategory.desktop';
+export { default as SkillCategoryMobile } from './SkillCategory.mobile';
