+++ src/features/skills/index.js
@@ -1,4 +1,5 @@
-import { useIsMobile } from '@hooks/useIsMobile';
+import { useIsMobile } from '@/hooks/useMedia';
+
 import SkillsDesktop from './Skills.desktop';
 import SkillsMobile from './Skills.mobile';
 
@@ -6,3 +7,5 @@ export default function Skills() {
   const isMobile = useIsMobile();
   return isMobile ? <SkillsMobile /> : <SkillsDesktop />;
 }
+
+export * from './components';
