+++ src/index.js
@@ -1,12 +1,11 @@
-// src/index.js
 import React from 'react';
-import ReactDOM from 'react-dom/client';
-import './styles/global.css';  // Make sure this path is correct
-import Portfolio from './Portfolio';
+import { createRoot } from 'react-dom/client';
+import App from './App';
+import './styles/global.css';
 
-const root = ReactDOM.createRoot(document.getElementById('root'));
+const root = createRoot(document.getElementById('root'));
 root.render(
   <React.StrictMode>
-    <Portfolio />
+    <App />
   </React.StrictMode>
 );
